Update readme for .org purposes
When we submit the plugin, the slug is determined by the name of the plugin and that can't be changed later, so we need it to be correct. 

I also updated the Plugin URI since this will not be listed at our Addons section, but instead on the Free Addons page and as a shortlink.